refactor(mvc)!: drop root sealing for the opt-in global contract
Sealing froze root-registered state read-only during server render to stop
cross-request bleed. It was both incomplete - shallow, so nested State, map,
has, and plain objects/arrays bled straight through - and harmful, freezing the
long-lived, mutable process globals (feature flags, config, an RPC surface) that
`static global` is meant to declare.

Opt-in global already prevents the accident sealing guarded: nothing reaches the
shared root now unless a class explicitly declares itself global. So sealing is
removed. `Context.sealing` becomes `Context.server` - a plain server-render flag
the adapter sets - used only to warn when a context-less non-global activates
(usually a missing per-request Provider). Request isolation comes from Provider;
a declared global is trusted to be process-wide and mutable.
