+++ packages/mvc/src/context.ts
@@ -1,5 +1,5 @@
 import { listener } from "./observable";
-import { event, seal, State, uid } from "./state";
+import { event, State, uid } from "./state";
 
 const LOOKUP = new WeakMap<State, Context>();
 let ROOT: Context;
@@ -20,11 +20,11 @@ declare namespace Context {
 
 class Context {
   /**
-   * When set, states registered to the root context are sealed read-only on
-   * activation. Adapters enable this during a server render, where the shared
-   * root is reused across requests and must not accept per-request mutation.
+   * Set by adapters during a server render (no DOM). Root-registered globals
+   * are shared across requests there, so a context-less non-global activated
+   * while this is on warns - it most often means a missing per-request Provider.
    */
-  static sealing = false;
+  static server = false;
 
   static get root(): Context {
     return ROOT ??= new Context();
@@ -234,8 +234,6 @@ class Context {
     const { cleanup, provide } = this;
     const root = this === Context.root;
     const TT: State.Extends[] = [];
-
-    if (root && Context.sealing) seal(I);
 
     function conflict(T: State.Extends) {
       const entries = provide.get(T);
